remove OpenRouter routing support from agent details and database schema
