+++ internal/store/sqlitestore/agents.go
@@ -29,14 +29,14 @@ func (s *SQLiteAgentStore) SetEmbeddingProvider(_ store.EmbeddingProvider) {}
 
 // agentSelectCols is the column list for all agent SELECT queries.
 const agentSelectCols = `id, agent_key, display_name, frontmatter, owner_id, provider, model,
-		 context_window, max_tool_iterations, workspace, restrict_to_workspace,
-		 tools_config, sandbox_config, subagents_config, memory_config,
-		 compaction_config, context_pruning, other_config,
-		 emoji, agent_description, thinking_level, max_tokens,
-		 self_evolve, skill_evolve, skill_nudge_interval,
-		 reasoning_config, workspace_sharing, chatgpt_oauth_routing,
-		 shell_deny_groups, kg_dedup_config, openrouter_routing,
-		 agent_type, is_default, status, budget_monthly_cents, created_at, updated_at, tenant_id`
+			 context_window, max_tool_iterations, workspace, restrict_to_workspace,
+			 tools_config, sandbox_config, subagents_config, memory_config,
+			 compaction_config, context_pruning, other_config,
+			 emoji, agent_description, thinking_level, max_tokens,
+			 self_evolve, skill_evolve, skill_nudge_interval,
+			 reasoning_config, workspace_sharing, chatgpt_oauth_routing,
+			 shell_deny_groups, kg_dedup_config,
+			 agent_type, is_default, status, budget_monthly_cents, created_at, updated_at, tenant_id`
 
 func (s *SQLiteAgentStore) Create(ctx context.Context, agent *store.AgentData) error {
 	if agent.ID == uuid.Nil {
@@ -51,15 +51,15 @@ func (s *SQLiteAgentStore) Create(ctx context.Context, agent *store.AgentData) e
 	}
 	_, err := s.db.ExecContext(ctx,
 		`INSERT INTO agents (id, agent_key, display_name, frontmatter, owner_id, provider, model,
-			 context_window, max_tool_iterations, workspace, restrict_to_workspace,
-			 tools_config, sandbox_config, subagents_config, memory_config,
-			 compaction_config, context_pruning, other_config,
-			 emoji, agent_description, thinking_level, max_tokens,
-			 self_evolve, skill_evolve, skill_nudge_interval,
-			 reasoning_config, workspace_sharing, chatgpt_oauth_routing,
-			 shell_deny_groups, kg_dedup_config, openrouter_routing,
-			 agent_type, is_default, status, budget_monthly_cents, created_at, updated_at, tenant_id)
-			 VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)`,
+				 context_window, max_tool_iterations, workspace, restrict_to_workspace,
+				 tools_config, sandbox_config, subagents_config, memory_config,
+				 compaction_config, context_pruning, other_config,
+				 emoji, agent_description, thinking_level, max_tokens,
+				 self_evolve, skill_evolve, skill_nudge_interval,
+				 reasoning_config, workspace_sharing, chatgpt_oauth_routing,
+				 shell_deny_groups, kg_dedup_config,
+				 agent_type, is_default, status, budget_monthly_cents, created_at, updated_at, tenant_id)
+				 VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)`,
 		agent.ID, agent.AgentKey,
 		agent.DisplayName,
 		sql.NullString{String: agent.Frontmatter, Valid: agent.Frontmatter != ""},
@@ -70,7 +70,7 @@ func (s *SQLiteAgentStore) Create(ctx context.Context, agent *store.AgentData) e
 		agent.Emoji, agent.AgentDescription, agent.ThinkingLevel, agent.MaxTokens,
 		agent.SelfEvolve, agent.SkillEvolve, agent.SkillNudgeInterval,
 		jsonOrEmpty(agent.ReasoningConfig), jsonOrEmpty(agent.WorkspaceSharing), jsonOrEmpty(agent.ChatGPTOAuthRouting),
-		jsonOrEmpty(agent.ShellDenyGroups), jsonOrEmpty(agent.KGDedupConfig), jsonOrEmpty(agent.OpenRouterRouting),
+		jsonOrEmpty(agent.ShellDenyGroups), jsonOrEmpty(agent.KGDedupConfig),
 		agent.AgentType, agent.IsDefault, agent.Status, agent.BudgetMonthlyCents,
 		now, now, tenantID,
 	)
@@ -151,7 +151,7 @@ func (s *SQLiteAgentStore) Update(ctx context.Context, id uuid.UUID, updates map
 		}
 	}
 	// NOT NULL JSON columns: null -> empty object.
-	for _, col := range []string{"other_config", "tools_config", "reasoning_config", "workspace_sharing", "chatgpt_oauth_routing", "shell_deny_groups", "kg_dedup_config", "openrouter_routing"} {
+	for _, col := range []string{"other_config", "tools_config", "reasoning_config", "workspace_sharing", "chatgpt_oauth_routing", "shell_deny_groups", "kg_dedup_config"} {
 		if v, ok := updates[col]; ok && v == nil {
 			updates[col] = []byte("{}")
 		}
@@ -257,7 +257,7 @@ func scanAgentRow(row agentRowScanner) (*store.AgentData, error) {
 	var d store.AgentData
 	var frontmatter sql.NullString
 	var toolsCfg, sandboxCfg, subagentsCfg, memoryCfg, compactionCfg, pruningCfg, otherCfg *[]byte
-	var reasoningCfg, wsCfg, oauthCfg, shellCfg, kgCfg, openrouterCfg *[]byte
+	var reasoningCfg, wsCfg, oauthCfg, shellCfg, kgCfg *[]byte
 	createdAt, updatedAt := scanTimePair()
 	err := row.Scan(
 		&d.ID, &d.AgentKey, &d.DisplayName, &frontmatter, &d.OwnerID, &d.Provider, &d.Model,
@@ -265,7 +265,7 @@ func scanAgentRow(row agentRowScanner) (*store.AgentData, error) {
 		&toolsCfg, &sandboxCfg, &subagentsCfg, &memoryCfg, &compactionCfg, &pruningCfg, &otherCfg,
 		&d.Emoji, &d.AgentDescription, &d.ThinkingLevel, &d.MaxTokens,
 		&d.SelfEvolve, &d.SkillEvolve, &d.SkillNudgeInterval,
-		&reasoningCfg, &wsCfg, &oauthCfg, &shellCfg, &kgCfg, &openrouterCfg,
+		&reasoningCfg, &wsCfg, &oauthCfg, &shellCfg, &kgCfg,
 		&d.AgentType, &d.IsDefault, &d.Status, &d.BudgetMonthlyCents,
 		createdAt, updatedAt, &d.TenantID,
 	)
@@ -312,9 +312,6 @@ func scanAgentRow(row agentRowScanner) (*store.AgentData, error) {
 	}
 	if kgCfg != nil {
 		d.KGDedupConfig = *kgCfg
-	}
-	if openrouterCfg != nil {
-		d.OpenRouterRouting = *openrouterCfg
 	}
 	return &d, nil
 }
