+++ ui/web/src/i18n/locales/en/providers.json
@@ -340,5 +340,37 @@
     "failedCreate": "Failed to create provider",
     "failedUpdate": "Failed to update provider",
     "failedDelete": "Failed to delete provider"
+  },
+  "routing": {
+    "sectionTitle": "Provider Routing",
+    "sectionDescription": "Configure how OpenRouter selects providers for your requests.",
+    "docsLink": "OpenRouter docs",
+    "sort": "Sort By",
+    "sortDefault": "Default (no preference)",
+    "sortPrice": "Price",
+    "sortThroughput": "Throughput",
+    "sortLatency": "Latency",
+    "sortHint": "Preferred sorting criterion for provider selection",
+    "allowFallbacks": "Allow Fallbacks",
+    "allowFallbacksHint": "Allow backup providers when primary is unavailable (default: enabled)",
+    "requireParameters": "Require Parameters",
+    "requireParametersHint": "Only use providers supporting all request parameters (default: disabled)",
+    "dataCollection": "Data Collection",
+    "dataCollectionDefault": "Default (provider choice)",
+    "dataCollectionAllow": "Allow",
+    "dataCollectionDeny": "Deny",
+    "dataCollectionHint": "Control whether providers may store request data for training",
+    "order": "Provider Order",
+    "orderHint": "Comma-separated provider slugs to try in order (e.g. anthropic, openai)",
+    "only": "Allowlist",
+    "onlyHint": "Only use these providers (comma-separated slugs, e.g. anthropic, openai)",
+    "ignore": "Skip List",
+    "ignoreHint": "Never use these providers (comma-separated slugs, e.g. together, deepinfra)",
+    "quantizations": "Quantizations",
+    "quantizationsHint": "Filter by quantization level (e.g. fp16, int8)",
+    "maxPrice": "Max Price",
+    "maxPricePrompt": "Prompt (per 1M tokens)",
+    "maxPriceCompletion": "Completion (per 1M tokens)",
+    "maxPriceHint": "Maximum price per million tokens. Leave empty for no limit."
   }
 }
