+++ ui/web/src/i18n/locales/vi/agents.json
@@ -172,8 +172,7 @@
       "instances": "Phiên bản người dùng",
       "permissions": "Phân quyền",
       "evolution": "Tiến Hóa",
-      "hooks": "Hooks",
-      "routing": "Định tuyến Provider"
+      "hooks": "Hooks"
     },
     "summonFailed": "Triệu hồi thất bại",
     "evolving": "Đang tiến hóa",
@@ -1098,37 +1097,5 @@
     "skillGrantFailed": "Không thể cấp skill",
     "skillRevoked": "Đã thu hồi skill",
     "skillRevokeFailed": "Không thể thu hồi skill"
-  },
-  "routing": {
-    "sectionTitle": "Định tuyến Provider",
-    "sectionDescription": "Cấu hình cách OpenRouter chọn provider cho các yêu cầu của bạn.",
-    "docsLink": "Tài liệu OpenRouter",
-    "sort": "Sắp xếp theo",
-    "sortDefault": "Mặc định (không ưu tiên)",
-    "sortPrice": "Giá",
-    "sortThroughput": "Thông lượng",
-    "sortLatency": "Độ trễ",
-    "sortHint": "Tiêu chí sắp xếp ưu tiên cho việc chọn provider",
-    "allowFallbacks": "Cho phép dự phòng",
-    "allowFallbacksHint": "Cho phép provider dự phòng khi provider chính không khả dụng (mặc định: bật)",
-    "requireParameters": "Yêu cầu tham số",
-    "requireParametersHint": "Chỉ sử dụng provider hỗ trợ tất cả tham số yêu cầu (mặc định: tắt)",
-    "dataCollection": "Thu thập dữ liệu",
-    "dataCollectionDefault": "Mặc định (theo provider)",
-    "dataCollectionAllow": "Cho phép",
-    "dataCollectionDeny": "Từ chối",
-    "dataCollectionHint": "Kiểm soát provider có thể lưu trữ dữ liệu yêu cầu để huấn luyện",
-    "order": "Thứ tự Provider",
-    "orderHint": "Slug provider phân tách bằng dấu phẩy theo thứ tự ưu tiên (vd: anthropic, openai)",
-    "only": "Danh sách cho phép",
-    "onlyHint": "Chỉ sử dụng các provider này (slug phân tách bằng dấu phẩy, vd: anthropic, openai)",
-    "ignore": "Danh sách bỏ qua",
-    "ignoreHint": "Không bao giờ sử dụng các provider này (slug phân tách bằng dấu phẩy, vd: together, deepinfra)",
-    "quantizations": "Lượng tử hóa",
-    "quantizationsHint": "Lọc theo mức lượng tử hóa (vd: fp16, int8)",
-    "maxPrice": "Giá tối đa",
-    "maxPricePrompt": "Prompt (trên 1M token)",
-    "maxPriceCompletion": "Completion (trên 1M token)",
-    "maxPriceHint": "Giá tối đa trên triệu token. Để trống nếu không giới hạn."
   }
 }
